% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotComp.R
\name{checkSeq}
\alias{checkSeq}
\title{validate the dot-bracket form sequences}
\usage{
checkSeq(seqs)
}
\arguments{
\item{seqs}{A string represent an RNA secondary structure in dot bracket form}
}
\value{
Returns an empty string when the sequence is validate, return an string "error"
if the sequence is not validate
}
\description{
A function that check if the sequence is validate, if the number of ( is not equal to
the number of ), then the motif sequence is not complete. When calling the dotComp
function will also run this funtion for validation.
}
\examples{
#this should return an empty string
checkSeq("...(((...)))...")

}
