% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotComp.R
\name{dotComp}
\alias{dotComp}
\title{Compute the difference between RNA secondary dot bracket form sequences}
\usage{
dotComp(seq1, seq2)
}
\arguments{
\item{seq1}{A string represent an RNA secondary structure in dot bracket form}

\item{seq2}{A string represent an RNA secondary structure in dot bracket form}
}
\value{
Returns a number represent the different in two structure
}
\description{
A function that calculates the difference between two dot bracket form of RNA sequences,
takes two sequences as function arguments.
}
\examples{
#this should return 2
dotComp("...(((...)))...", ".(.(.(...)))...")


}
