% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dotCompExample}
\alias{dotCompExample}
\title{Result from miRNA Motif Analysis}
\format{
A matrix with columns:
\describe{
\item{id}{name for the miRNA}
\item{sequences}{miRNA sequence in dot-bracket form}
}
}
\source{
\url{https://github.com/Deemolotus/BCB330Y-and-BCB430Y/tree/master/Main}
}
\usage{
dotCompExample
}
\description{
A dataset contains normal and mutated miRNA secondary structure
}
\examples{
\dontrun{
 dotCompExample
}
}
\keyword{datasets}
