% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleSequence}
\alias{ExampleSequence}
\title{Part of miRNA sequence data from miRBase}
\format{
A matrix with columns:
\describe{
\item{id}{name for the miRNA}
\item{sequences}{miRNA sequence}
}
}
\source{
\url{http://www.mirbase.org/}
}
\usage{
ExampleSequence
}
\description{
A dataset contains 14 different miRNA
}
\examples{
\dontrun{
 ExampleSequence
}
}
\keyword{datasets}
