% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plotter}
\alias{plotter}
\title{Plotting the differences}
\usage{
plotter(dataframes)
}
\arguments{
\item{dataframes}{A list that contains all id, sequences, differences.}
}
\value{
Returns a histogram.
}
\description{
A function that takes in a list of dataframes and convert the
numerical number to plot form for easier interpretation,
only use this function when there is a reference sequence,
if the function used is comparison in pair,
plotting them in this way is meaning less
}
\examples{
# Examples 1:
# plot the difference
a <- list()
x <- data.frame(id = c("1","2","3","4","5","6","7","8"),
                sequences = c("a", "b", "c", "d", "e", "f", "g", "h"),
                difference = c(1,5,3,4,5,6,6,32))
a <- x
plotter(a)

# Examples 2:
# connect this function to seqCompareAsInput.R

library(gscVisualizer)
filePath <- system.file("extdata", "test.fa", package = "gscVisualizer")
a <- seqCompareAsFile("RNA", "noncoding", filePath)
plotter(a)

}
