% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runGscVisualizer.R
\name{runGscVisualizer}
\alias{runGscVisualizer}
\title{Launch Shiny App For Package gscVisualizer}
\usage{
runGscVisualizer()
}
\value{
No return value but open up a shiny page.
}
\description{
A function that launches the shiny app for this package.
The shiny app allows user to choose a input file, the app
will calculate the difference among those genes and generate
a plot shows the distribution for all differences.
}
\examples{
\dontrun{
runGscVisualizer()
}

}
\references{
Steipe B., ABC project (.utility 4.07) A Bioinformatics Course: Applied Bioinformatics
\href{http://steipe.biochemistry.utoronto.ca/abc/index.php/Bioinformatics_Main_Page}{Link}.
}
\author{
Zhiwen Tan, \email{leon.tan@mail.utoronto.ca}
}
