% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqCompareAsInput.R
\name{seqCompareAsFile}
\alias{seqCompareAsFile}
\title{Calculate the difference between gene sequences}
\usage{
seqCompareAsFile(seqType, type, filePath)
}
\arguments{
\item{seqType}{The sequence ID for refernce sequence.}

\item{type}{A string indicating either the comparison is for coding gene,
or noncoding gene. type = "coding" or type = "noncoding".}

\item{filePath}{A string shows the path for the specific .fa file}
}
\value{
Returns an List object of class seqCompareAsInput with results.
\itemize{
\item id - A string name for the correspond sequence.
\item sequences - A string indicate the sequence.
\item difference - A value of class "numeric" indicating the
difference between this sequence and reference sequence.
}
}
\description{
A function that calculates the difference between gene sequences, takes
an .fa file as function arguments, the default reference is the first sequence in the .fa file
the function returns a list shows all differences. For coding genes,
the difference are calculated based on the amino acid sequence.
For noncoding gene, the difference are calculated based on Needleman–Wunsch algorithm.
}
\examples{
# Examples 1:
# Using a .fa file for miRNA, use this file to
# Calculate the difference
seqCompareAsFile("RNA", "noncoding", "../../inst/extdata/test.fa")

# Examples 2:
# Using a .fa file for DNA sequence, use this file to
# Calculate the difference
seqCompareAsFile("DNA", "coding", "../../inst/extdata/seqCompareAsFileTest.fa")

}
\references{
Charif D, Lobry J (2007). “SeqinR 1.0-2: a contributed package to the R project for statistical
computing devoted to biological sequences retrieval and analysis.” In Bastolla U, Porto M, Roman H,
Vendruscolo M (eds.), \emph{Structural approaches to sequence evolution: Molecules, networks, populations,
series Biological and Medical Physics, Biomedical Engineering}, 207-232.
Springer Verlag, New York. ISBN : 978-3-540-35305-8.

Kevin R. Coombes (2020). NameNeedle: Using Needleman-Wunsch to Match Sample Names. R package version
1.2.6/r51. \href{https://R-Forge.R-project.org/projects/nameneedle/}{Link}

Pagès H, Aboyoun P, Gentleman R, DebRoy S (2020). Biostrings: Efficient manipulation of biological strings.
R package version 2.58.0, \href{https://bioconductor.org/packages/Biostrings}{Link}.
}
