% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqCompareAsInput.R
\name{seqCompareAsInput}
\alias{seqCompareAsInput}
\title{Calculate the difference between gene sequences}
\usage{
seqCompareAsInput(ID, ref, type, ...)
}
\arguments{
\item{ID}{The sequence ID for refernce sequence.}

\item{ref}{A string for the reference sequence.}

\item{type}{A string indicating either the comparison is for coding gene,
or noncoding gene. type = "coding" or type = "noncoding".}

\item{...}{A list of mutated sequence string where the odd index indicate the mutated sequence ID,
the even index indicate the mutated sequence.}
}
\value{
Returns an List object of class seqCompareAsInput with results.
\itemize{
\item id - A string name for the correspond sequence.
\item sequences - A string indicate the sequence.
\item difference - A value of class "numeric" indicating the
difference between this sequence and reference sequence.
}
}
\description{
A function that calculates the difference between gene sequences, takes
sequences as function arguments, every sequences are compare to the reference sequence,
eg. if the input is seqCompareAsInput(A,B,C,D,E,F) where A is the reference sequence,
then B,C,D,E,F are compare with sequence A.
the function returns a list shows all differences. For coding genes,
the difference are calculated based on the amino acid sequence.
For noncoding gene, the difference are calculated based on Needleman–Wunsch algorithm.
}
\examples{
# Examples 1:
# Using random sequences for noncoding type
# Calculate the difference
seqCompareAsInput(ID = "hsa-let-7b",
                  ref = "CGGGGUGAGGUAGUAGGUUGUGUGGUUUCAGGGCAGUGAUGU
                         UGCCCCUCGGAAGAUAACUAUACAACCUACUGCCUUCCCUG",
                  type = "noncoding",
                  "hsa-let-7b_UCEC",
                  "CGGGGUGAGGUAGUAGGUUGUAUGGUUUCAGGGCAGUGAUGU
                   UGCCCCUCGGAAGAUAACUAUACAACCUACUGCCUUCCCUG",
                  "hsa-let-7b_LIRI",
                  "CGGGGUGAGGUAGUAGGUUGUGUGGUCUCAGGGCAGUGAUGU
                   UGCCCCUCGGAAGAUAACUAUACAACCUACUGCCUUCCCUG")

# Examples 2:
# Using random sequences for coding type
# Calculate the difference
filePath <- system.file("extdata", "seqCompareAsInputExample.rds", package = "gscVisualizer")
#readfrom rds file
seqs <- readRDS(filePath)
seqCompareAsInput(ID = seqs[1],
                  ref = seqs[2],
                 type = seqs[3],
                 seqs[4],
                 seqs[5])

}
\references{
Charif D, Lobry J (2007). “SeqinR 1.0-2: a contributed package to the R project for statistical
computing devoted to biological sequences retrieval and analysis.” In Bastolla U, Porto M, Roman H,
Vendruscolo M (eds.), \emph{Structural approaches to sequence evolution: Molecules, networks, populations,
series Biological and Medical Physics, Biomedical Engineering}, 207-232.
Springer Verlag, New York. ISBN : 978-3-540-35305-8.

Kevin R. Coombes (2020). NameNeedle: Using Needleman-Wunsch to Match Sample Names. R package version
1.2.6/r51. \href{https://R-Forge.R-project.org/projects/nameneedle/}{Link}
}
