% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seqCompareAsInput.R
\name{seqCompareInPair}
\alias{seqCompareInPair}
\title{Calculate the difference between each pair of gene sequences}
\usage{
seqCompareInPair(type, ...)
}
\arguments{
\item{type}{A string indicating either the comparison is for coding gene,
or noncoding gene. type = "coding" or type = "noncoding".}

\item{...}{A list of sequence string, comparing the difference between each pair of sequence.
For example, first input compare with second sequence,
third one compare to the forth one and so on.}
}
\value{
Returns an List object of class seqCompareAsInput with results.
\itemize{
\item id - A string name for the correspond sequence.
\item sequences - A string indicate the sequence.
\item difference - A value of class "numeric" indicating the
difference between this sequence and reference sequence.
}
}
\description{
A function that calculates the difference between gene sequences, takes
sequences as function arguments, compare the difference in pairs of sequences.
eg. if the input is seqCompareAsInput(A,B,C,D,E,F),
then sequence A will compare with sequence B, sequence C compare to sequence D, and so on.
the function returns a list shows all differences. For coding genes,
the difference are calculated based on the amino acid sequence.
For noncoding gene, the difference are calculated based on Needleman–Wunsch algorithm.
}
\examples{
seqCompareInPair("noncoding", "seq1", "ATCGTCGCC",
                 "seq2", "TTCGTCGCG", "seq3", "ATCTTGGCC", "seq4", "ATGGTAGGG")

}
\references{
Charif D, Lobry J (2007). “SeqinR 1.0-2: a contributed package to the R project for statistical
computing devoted to biological sequences retrieval and analysis.” In Bastolla U, Porto M, Roman H,
Vendruscolo M (eds.), \emph{Structural approaches to sequence evolution: Molecules, networks, populations,
series Biological and Medical Physics, Biomedical Engineering}, 207-232.
Springer Verlag, New York. ISBN : 978-3-540-35305-8.

Kevin R. Coombes (2020). NameNeedle: Using Needleman-Wunsch to Match Sample Names. R package version
1.2.6/r51. \href{https://R-Forge.R-project.org/projects/nameneedle/}{Link}
}
