% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotComp.R
\name{sequenceHandle}
\alias{sequenceHandle}
\title{Compute the difference between RNA secondary dot bracket form sequences when
the two sequences are not in the same length}
\usage{
sequenceHandle(seq1, seq2)
}
\arguments{
\item{seq1}{A string represent an RNA secondary structure in dot bracket form}

\item{seq2}{A string vector represent an RNA secondary structure in dot bracket form}
}
\value{
Returns a number represent the different in two structure
}
\description{
A function that calculates the difference between two dot bracket form of RNA sequences,
takes two sequences as function arguments. this function handle the case when two
sequences have different length, if you are computing difference for a deletion or
insertion mutation, you can use this function directly
}
\examples{
#this example should return 4
sequenceHandle("((..))", "(....)..")

}
