% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dotComp.R
\name{dotComp}
\alias{dotComp}
\title{Compute the difference between RNA secondary dot bracket form sequences}
\usage{
dotComp(seq1, seq2, option = 0)
}
\arguments{
\item{seq1}{A string represent an RNA secondary structure in dot bracket form}

\item{seq2}{A string represent an RNA secondary structure in dot bracket form}

\item{option}{Anumber represent the method for comparison, if option is 0, then
the function will compare every characters in the two sequence, if
option is 1, then the function will use string alignment method
(Needleman–Wunsch algorithm) for comparison}
}
\value{
Returns a number represent the different in two structure
}
\description{
A function that calculates the difference between two dot bracket form of RNA sequences,
takes two sequences as function arguments.This function can handle point mutation,
insertion, and deletion.
}
\examples{
#this should return 2
dotComp("...(((...)))...", ".(.(.(...)))...")


}
